@@ -1,6 +1,6 @@
 name: shared_preferences_linux
 description: Linux implementation of the shared_preferences plugin
-version: 0.0.3+1
+version: 0.0.4-nullsafety
 homepage: https://github.com/flutter/plugins/tree/master/packages/shared_preferences/shared_preferences_linux
 
 flutter:
@@ -11,17 +11,16 @@ flutter:
         pluginClass: none
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0-259.8.beta <3.0.0"
   flutter: ">=1.12.8"
 
 dependencies:
-  file: ">=5.1.0 <7.0.0"
+  file: ">=6.0.0-nullsafety.4 <7.0.0"
   flutter:
     sdk: flutter
   meta: ^1.0.4
-  path: ^1.6.4
-  path_provider_linux: ^0.0.1
-  shared_preferences_platform_interface: ^1.0.0
+  path_provider_linux: ^0.2.0-nullsafety
+  shared_preferences_platform_interface: ^2.0.0-nullsafety
 
 dev_dependencies:
   flutter_test:
