@@ -1,7 +1,7 @@
 name: shared_preferences_windows
 description: Windows implementation of shared_preferences
 homepage: https://github.com/flutter/plugins/tree/master/packages/shared_preferences/shared_preferences_windows
-version: 0.0.2+3
+version: 0.0.3-nullsafety
 
 flutter:
   plugin:
@@ -11,18 +11,18 @@ flutter:
         pluginClass: none
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0-259.8.beta <3.0.0"
   flutter: ">=1.12.8"
 
 dependencies:
-  shared_preferences_platform_interface: ^1.0.0
+  shared_preferences_platform_interface: ^2.0.0-nullsafety
   flutter:
     sdk: flutter
-  file: ">=5.1.0 <7.0.0"
+  file: ">=6.0.0-nullsafety.4 <7.0.0"
   meta: ^1.1.7
   path: ^1.6.4
-  path_provider_platform_interface: ^1.0.3
-  path_provider_windows: ^0.0.2
+  path_provider_platform_interface: ^2.0.0-nullsafety
+  path_provider_windows: ^0.1.0-nullsafety.2
 
 dev_dependencies:
   flutter_test:
