@@ -20,24 +20,24 @@ class SharedPreferencesLinux extends SharedPreferencesStorePlatform {
   static SharedPreferencesLinux instance = SharedPreferencesLinux();
 
   /// Local copy of preferences
-  Map<String, Object> _cachedPreferences;
+  Map<String, Object>? _cachedPreferences;
 
   /// File system used to store to disk. Exposed for testing only.
   @visibleForTesting
-  FileSystem fs = LocalFileSystem();
+  FileSystem? fs = LocalFileSystem();
 
   /// Gets the file where the preferences are stored.
   Future<File> _getLocalDataFile() async {
     final pathProvider = PathProviderLinux();
     final directory = await pathProvider.getApplicationSupportPath();
-    return fs.file(path.join(directory, 'shared_preferences.json'));
+    return fs!.file(path.join(directory!, 'shared_preferences.json'));
   }
 
   /// Gets the preferences from the stored file. Once read, the preferences are
   /// maintained in memory.
   Future<Map<String, Object>> _readPreferences() async {
     if (_cachedPreferences != null) {
-      return _cachedPreferences;
+      return _cachedPreferences!;
     }
 
     _cachedPreferences = {};
@@ -45,11 +45,11 @@ class SharedPreferencesLinux extends SharedPreferencesStorePlatform {
     if (localDataFile.existsSync()) {
       String stringMap = localDataFile.readAsStringSync();
       if (stringMap.isNotEmpty) {
-        _cachedPreferences = json.decode(stringMap) as Map<String, Object>;
+        _cachedPreferences = json.decode(stringMap) as Map<String, Object>?;
       }
     }
 
-    return _cachedPreferences;
+    return _cachedPreferences!;
   }
 
   /// Writes the cached preferences to disk. Returns [true] if the operation
