@@ -7,7 +7,7 @@ import 'package:path/path.dart' as path;
 import 'package:path_provider_linux/path_provider_linux.dart';
 import 'package:shared_preferences_linux/shared_preferences_linux.dart';
 
-MemoryFileSystem fs;
+MemoryFileSystem? fs;
 
 void main() {
   setUp(() {
@@ -19,17 +19,17 @@ void main() {
   Future<String> _getFilePath() async {
     final pathProvider = PathProviderLinux();
     final directory = await pathProvider.getApplicationSupportPath();
-    return path.join(directory, 'shared_preferences.json');
+    return path.join(directory!, 'shared_preferences.json');
   }
 
   _writeTestFile(String value) async {
-    fs.file(await _getFilePath())
+    fs!.file(await _getFilePath())
       ..createSync(recursive: true)
       ..writeAsStringSync(value);
   }
 
   Future<String> _readTestFile() async {
-    return fs.file(await _getFilePath()).readAsStringSync();
+    return fs!.file(await _getFilePath()).readAsStringSync();
   }
 
   SharedPreferencesLinux _getPreferences() {
